+++ kroxylicious-api/src/main/java/io/kroxylicious/proxy/filter/RequestFilter.java
@@ -25,7 +25,7 @@ public interface RequestFilter extends Filter {
 
     /**
      * Does this filter implementation want to handle a request. If so, the
-     * {@link #onRequest(ApiKeys, RequestHeaderData, ApiMessage, FilterContext)} method
+     * {@link #onRequest(ApiKeys, short, RequestHeaderData, ApiMessage, FilterContext)} method
      * will be eligible to be called with the deserialized request data (if the
      * message reaches this filter in the filter chain).
      *
@@ -52,9 +52,42 @@ public interface RequestFilter extends Filter {
      *         request to be forwarded.
      * @see io.kroxylicious.proxy.filter Creating Filter Result objects
      * @see io.kroxylicious.proxy.filter Thread Safety
+     * @deprecated implement {@link #onRequest(ApiKeys, short, RequestHeaderData, ApiMessage, FilterContext)} instead.
      */
-    CompletionStage<RequestFilterResult> onRequest(ApiKeys apiKey,
-                                                   RequestHeaderData header,
-                                                   ApiMessage request,
-                                                   FilterContext context);
+    @Deprecated(forRemoval = true, since = "0.19.0")
+    default CompletionStage<RequestFilterResult> onRequest(ApiKeys apiKey,
+                                                           RequestHeaderData header,
+                                                           ApiMessage request,
+                                                           FilterContext context) {
+        throw new UnsupportedOperationException("implement #onRequest(ApiKeys, short, RequestHeaderData, ApiMessage, FilterContext)");
+    }
+
+    /**
+     * Handle the given {@code header} and {@code request} pair, returning the {@code header} and {@code request}
+     * pair to be passed to the next filter using the RequestFilterResult.
+     * <br/>
+     * The implementation may modify the given {@code header} and {@code request} in-place, or instantiate a
+     * new instances.
+     *
+     * @param apiKey key of the request
+     * @param apiVersion api version of the request
+     * @param header header of the request
+     * @param request body of the request
+     * @param context context containing methods to continue the filter chain and other contextual data
+     * @return a non-null CompletionStage that, when complete, will yield a RequestFilterResult containing the
+     *         request to be forwarded.
+     * @see io.kroxylicious.proxy.filter Creating Filter Result objects
+     * @see io.kroxylicious.proxy.filter Thread Safety
+     */
+    @SuppressWarnings("deprecated")
+    default CompletionStage<RequestFilterResult> onRequest(ApiKeys apiKey,
+                                                           short apiVersion,
+                                                           RequestHeaderData header,
+                                                           ApiMessage request,
+                                                           FilterContext context) {
+        // default implementation exists so that pre-0.19.0 implementations of RequestFilter continue to work without change.
+        // when the deprecated method is removed, remove this default implementation.
+        return onRequest(apiKey, header, request, context);
+    }
+
 }