+++ kroxylicious-api/src/main/java/io/kroxylicious/proxy/filter/ResponseFilter.java
@@ -25,7 +25,7 @@ public interface ResponseFilter extends Filter {
 
     /**
      * Does this filter implementation want to handle a response. If so, the
-     * {@link #onResponse(ApiKeys, ResponseHeaderData, ApiMessage, FilterContext)} method
+     * {@link #onResponse(ApiKeys, short, ResponseHeaderData, ApiMessage, FilterContext)} method
      * will be eligible to be called with the deserialized response data (if the
      * message reaches this filter in the filter chain).
      *
@@ -44,7 +44,7 @@ public interface ResponseFilter extends Filter {
      * The implementation may modify the given {@code header} and {@code response} in-place, or instantiate a
      * new instances.
      *
-     * @param apiKey   key of the request
+     * @param apiKey   key of the response
      * @param header   response header.
      * @param response The body to handle.
      * @param context  The context.
@@ -52,9 +52,41 @@ public interface ResponseFilter extends Filter {
      *         response to be forwarded.
      * @see io.kroxylicious.proxy.filter Creating Filter Result objects
      * @see io.kroxylicious.proxy.filter Thread Safety
+     * @deprecated implement {@link #onResponse(ApiKeys, short, ResponseHeaderData, ApiMessage, FilterContext)} instead.
      */
-    CompletionStage<ResponseFilterResult> onResponse(ApiKeys apiKey,
-                                                     ResponseHeaderData header,
-                                                     ApiMessage response,
-                                                     FilterContext context);
+    @Deprecated(forRemoval = true, since = "0.19.0")
+    default CompletionStage<ResponseFilterResult> onResponse(ApiKeys apiKey,
+                                                             ResponseHeaderData header,
+                                                             ApiMessage response,
+                                                             FilterContext context) {
+        throw new UnsupportedOperationException("implement #onResponse(ApiKeys, short, ResponseHeaderData, ApiMessage, FilterContext)");
+    }
+
+    /**
+     * Handle the given {@code header} and {@code response} pair, returning the {@code header} and {@code response}
+     * pair to be passed to the next filter using the ResponseFilterResult.
+     * <br/>
+     * The implementation may modify the given {@code header} and {@code response} in-place, or instantiate a
+     * new instances.
+     *
+     * @param apiKey key of the response
+     * @param apiVersion api version of the response
+     * @param header response header.
+     * @param response The body to handle.
+     * @param context The context.
+     * @return a non-null CompletionStage that, when complete, will yield a ResponseFilterResult containing the
+     *         response to be forwarded.
+     * @see io.kroxylicious.proxy.filter Creating Filter Result objects
+     * @see io.kroxylicious.proxy.filter Thread Safety
+     */
+    @SuppressWarnings("deprecated")
+    default CompletionStage<ResponseFilterResult> onResponse(ApiKeys apiKey,
+                                                             short apiVersion,
+                                                             ResponseHeaderData header,
+                                                             ApiMessage response,
+                                                             FilterContext context) {
+        // default implementation exists so that pre-0.19.0 implementations of ResponseFilter continue to work without change.
+        // when the deprecated method is removed, remove this default implementation.
+        return onResponse(apiKey, header, response, context);
+    }
 }